# src/main/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
  "/user/notifications":
    get:
      tags:
        - User
      summary: Получить настройки уведомлений
      operationId: userNotificationsGet
      responses:
        '200':
          description: Настройки уведомлений
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/NotificationPreferences"
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
    put:
      tags:
        - User
      summary: Обновить настройки уведомлений
      operationId: userNotificationsPut
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/NotificationPreferences"
      responses:
        '200':
          description: Настройки обновлены
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
        '404':
          description: Настройки уведомлений не найдены
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
    delete:
      tags:
        - User
      summary: Отменить уведомления
      operationId: userNotificationsDelete
      responses:
        '204':
          description: Настройки удалены
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
...
  "/user/telegram/link":
    post:
      tags:
        - User
      summary: Инициировать привязку Telegram
      operationId: userTelegramLinkPost
      responses:
        '200':
          description: Код для /start команды в боте
          content:
            text/plain:
              schema:
                type: string
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/ErrorResponse"
...
components:
  schemas:
    ErrorDetail:
      type: object
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: Должен быть валидным email адресом
      description: Детали ошибки (опционально)
    ErrorResponse:
      required:
        - code
        - level
        - message
      type: object
      properties:
        code:
          type: string
          description: Машинно-читаемый код ошибки
          example: VALIDATION_FAILED
        level:
          type: string
          description: Уровень серьезности
          example: error
          enum:
            - ErrorResponse.LevelEnum.INFO(value=info)
            - ErrorResponse.LevelEnum.WARNING(value=warning)
            - ErrorResponse.LevelEnum.ERROR(value=error)
        message:
          type: string
          description: Человеко-читаемое сообщение
          example: Некоторые поля заполнены неверно
        details:
          type: array
          description: Детали ошибки (опционально)
          items:
            "$ref": "#/components/schemas/ErrorDetail"
    NotificationPreferences:
      type: object
      properties:
        notifyNewEvents:
          type: boolean
        notifyUpcoming:
          type: boolean
        notifyBeforeHours:
          maximum: 24
          minimum: 1
          type: integer
          format: int32
...